+++ drizzle/0008_show_crowd.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "user" ADD COLUMN "show_crowd" boolean;